* fix the animation timing, the notification `start_time` must be in millisec to avoid losing precision * convert the animation duration to millisec once during settings parsing * decrease the notification margin from 10.0 to 5.0, it looked way shifted on larger screens * pass around instances on std::chrono instead of float, use float in the final stage when processing/operating-on the time (if needed) * decrease the animation duration to 0.35 sec, looks more swishy!

+++ dll/dll/settings.h
@@ -124,8 +124,9 @@ struct Overlay_Appearance {
     float notification_g = 0.29f;
     float notification_b = 0.48f;
     float notification_a = 1.0f;
-    float notification_rounding = 0.0f;
-    float notification_animation = 0.0f;
+    float notification_rounding = 0.0f; // corners roundness for all notifications
+    uint32 notification_animation = 0; // sliding animation duration (millisec)
+
     std::string ach_unlock_datetime_format = "%Y/%m/%d - %H:%M:%S";
     
     float background_r = -1.0f;
